# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис сокращения ссылок",
        "title": "URL SHORTNER API",
        "contact": {
            "email": "t1mon.ggg@yandex.ru"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Ping"
                ],
                "summary": "Запрос состояния хранилища",
                "responses": {
                    "200": {
                        "description": "Проверка подключения к хранилищу успешно завершено"
                    },
                    "500": {
                        "description": "Хранилище недоступно"
                    }
                }
            },
            "post": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Create"
                ],
                "summary": "Запрос на сокращение ссылки",
                "parameters": [
                    {
                        "description": "Сокращаемый URL",
                        "name": "Input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Идентификационный cookie Client_ID",
                        "name": "Client_ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Создана новая сокращенная ссылка",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Запрашиваемый URL уже существует",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APICreate"
                ],
                "summary": "Запрос на сокращение ссылки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификационный cookie Client_ID",
                        "name": "Client_ID",
                        "in": "header"
                    },
                    {
                        "description": "Сокращаемый URL",
                        "name": "Input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webhandlers.lURL"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Создана новая сокращенная ссылка",
                        "schema": {
                            "$ref": "#/definitions/webhandlers.sURL"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос"
                    },
                    "409": {
                        "description": "Запрашиваемый URL уже существует",
                        "schema": {
                            "$ref": "#/definitions/webhandlers.sURL"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/api/shorten/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APICreate"
                ],
                "summary": "Запрос на сокращение ссылок списком",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификационный cookie Client_ID",
                        "name": "Client_ID",
                        "in": "header"
                    },
                    {
                        "description": "Список сокращаемых URLs",
                        "name": "Input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webhandlers.input"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Список успешно обработан",
                        "schema": {
                            "$ref": "#/definitions/webhandlers.output"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/api/user/urls": {
            "get": {
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ListAll"
                ],
                "summary": "Запрос на получение всех сокращенных ссылок пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификационный cookie Client_ID",
                        "name": "Client_ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сокращений пользователя",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/webhandlers.answer"
                            }
                        }
                    },
                    "202": {
                        "description": "Сокращения в базе данных отсутствуют"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "APIDelete"
                ],
                "summary": "Запрос на удаление короткой ссылки",
                "parameters": [
                    {
                        "description": "Список удаляемых коротких идентификаторов",
                        "name": "Input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Идентификационный cookie Client_ID",
                        "name": "Client_ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Запрос принят в обработку"
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        }
    },
    "definitions": {
        "webhandlers.answer": {
            "type": "object",
            "properties": {
                "original_url": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                }
            }
        },
        "webhandlers.input": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                }
            }
        },
        "webhandlers.lURL": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "webhandlers.output": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                }
            }
        },
        "webhandlers.sURL": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        }
    }
}